Revert "Convert double to float"
This reverts commit 12c749d7a3a656c99ae840e12551bd1b59902d71.

+++ common/src/main/java/dk/itu/common/models/WithBoundingBox.java
@@ -1,8 +1,8 @@
 package dk.itu.common.models;
 
 public interface WithBoundingBox {
-    double minLon();
-    double minLat();
-    double maxLon();
-    double maxLat();
+    float minLon();
+    float minLat();
+    float maxLon();
+    float maxLat();
 }
